+++ cirq-core/cirq/sim/clifford/stabilizer_simulation_state.py
@@ -22,7 +22,7 @@ from cirq._compat import deprecated_parameter
 from cirq.ops import common_gates, global_phase_op, matrix_gates, swap_gates
 from cirq.ops.clifford_gate import SingleQubitCliffordGate
 from cirq.protocols import has_unitary, num_qubits, unitary
-from cirq.sim.act_on_args import DenseSimulationState
+from cirq.sim.dense_simulation_state import DenseSimulationState
 from cirq.type_workarounds import NotImplementedType
 
 if TYPE_CHECKING:
