+++ cirq-core/cirq/sim/clifford/stabilizer_simulation_state_test.py
