+++ cirq-core/cirq/testing/consistent_act_on.py
@@ -22,11 +22,11 @@ from cirq.ops import common_gates
 from cirq.ops.dense_pauli_string import DensePauliString
 from cirq import protocols
 from cirq.qis import clifford_tableau
-from cirq.sim import act_on_state_vector_args, final_state_vector
+from cirq.sim import state_vector_simulation_state, final_state_vector
 from cirq.sim.clifford import (
-    act_on_clifford_tableau_args,
+    clifford_tableau_simulation_state,
     stabilizer_state_ch_form,
-    act_on_stabilizer_ch_form_args,
+    ch_form_simulation_state,
 )
 
 
@@ -46,7 +46,7 @@ def state_vector_has_stabilizer(state_vector: np.ndarray, stabilizer: DensePauli
     """
 
     qubits = LineQubit.range(protocols.num_qubits(stabilizer))
-    args = act_on_state_vector_args.StateVectorSimulationState(
+    args = state_vector_simulation_state.StateVectorSimulationState(
         available_buffer=np.empty_like(state_vector),
         qubits=qubits,
         prng=np.random.RandomState(),
@@ -159,7 +159,7 @@ def _final_clifford_tableau(
         the tableau otherwise."""
 
     tableau = clifford_tableau.CliffordTableau(len(qubit_map))
-    args = act_on_clifford_tableau_args.CliffordTableauSimulationState(
+    args = clifford_tableau_simulation_state.CliffordTableauSimulationState(
         tableau=tableau, qubits=list(qubit_map.keys()), prng=np.random.RandomState()
     )
     for op in circuit.all_operations():
@@ -187,7 +187,7 @@ def _final_stabilizer_state_ch_form(
         returns the StabilizerStateChForm otherwise."""
 
     stabilizer_ch_form = stabilizer_state_ch_form.StabilizerStateChForm(len(qubit_map))
-    args = act_on_stabilizer_ch_form_args.CHFormSimulationState(
+    args = ch_form_simulation_state.CHFormSimulationState(
         qubits=list(qubit_map.keys()),
         prng=np.random.RandomState(),
         initial_state=stabilizer_ch_form,
